Move the constant folding logic into a separate utility which can be used by multiple passes.
NFC
